Modularized the backend more. Logic split now for both ML and OSM operations outside of ParkApi 'business logic'... TODO -- implement SQLAlchemy... implemented threading
